Move prototype data to `FeedImageViewModel+PrototypeData` extension
